PH-275: Create DQI evaluations launcher service
